Maintenance: Reorganized code for better structure
